Refactor UI and add Testimonials component
Several changes were made to the UI and the build of the application. The most significant change is the addition of a Testimonials component which allows to display customer feedback. Other changes include updating the layout of the about me section and making adjustments to CSS to refine the look of the application. Updating the deployment.xml file paths to reflect the modifications.