Refactor styling in Layout and Update Freelancer Status
This commit makes two primary changes. Firstly, it refactors the styling on the "App.tsx" and "main.scss", especially on the logo & personal picture, making it responsive for better scalability and readability. Secondly, it alters the status of the freelancer from "Top Rated Freelancer" to "100% Job Success Score" in "AboutMe.tsx", providing a more accurate representation of current status.

Took 21 minutes